+++ src/main/kotlin/ru/tinkoff/fintech/service/cashback/CashbackRule.kt
@@ -1,7 +1,6 @@
 package ru.tinkoff.fintech.service.cashback
 
 import ru.tinkoff.fintech.model.TransactionInfo
-import java.time.Month
 
 interface CashbackRule {
