+++ src/main/kotlin/ru/tinkoff/fintech/service/cashback/SixSixSixRule.kt
@@ -2,7 +2,7 @@ package ru.tinkoff.fintech.service.cashback
 
 import ru.tinkoff.fintech.model.TransactionInfo
 
-class SixSixSix : CashbackRule {
+class SixSixSixRule : CashbackRule {
     override fun calculate(transactionInfo: TransactionInfo) =
         if (transactionInfo.transactionSum % 666.0 == .0)  6.66 else .0
 }
